+++ install.sh
@@ -1995,7 +1995,7 @@ webhooks:
           - CREATE
           - UPDATE
         resources:
-          - schedule
+          - schedules
   - clientConfig:
       caBundle: "${CA_BUNDLE}"
       service:
@@ -2016,7 +2016,7 @@ webhooks:
           - CREATE
           - UPDATE
         resources:
-          - workflow
+          - workflows
   - clientConfig:
       caBundle: "${CA_BUNDLE}"
       service: